requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.50.0
python-dotenv==1.0.0
python-multipart>=0.0.20
httpx>=0.27.0
